+++ packages/application-shell/src/components/application-shell-authenticated/new-main-navigation-flag-wrapper.tsx
@@ -1,12 +1,25 @@
 // TODO: Remove this file once the new navigation is enabled for all users
-import { useFeatureToggle } from '@flopflip/react-broadcast';
+import { useFeatureToggle, useAdapterStatus } from '@flopflip/react-broadcast';
 import { MAIN_NAVIGATION } from '../../feature-toggles';
 
+type TNewMainNavigationFlagWrapperChildrenProps = {
+  isNewNavigationEnabled: boolean;
+  isNewNavigationEnabledEvaluationPending: boolean;
+};
+
 const NewMainNavigationFlagWrapper = (props: {
-  children: (isNewNavigationEnabled: boolean) => JSX.Element;
+  children: ({
+    isNewNavigationEnabled,
+    isNewNavigationEnabledEvaluationPending,
+  }: TNewMainNavigationFlagWrapperChildrenProps) => JSX.Element;
 }) => {
   const isNewNavigationEnabled = useFeatureToggle(MAIN_NAVIGATION);
-  return props.children(isNewNavigationEnabled);
+  const { isConfiguring } = useAdapterStatus();
+
+  return props.children({
+    isNewNavigationEnabled,
+    isNewNavigationEnabledEvaluationPending: isConfiguring,
+  });
 };
 
 export default NewMainNavigationFlagWrapper;
